Externalizing TableProps
