+++ kie-wb-common-react/boxed-expression-editor/src/api/Table.ts
@@ -15,6 +15,47 @@
  */
 
 import { DataType } from "./DataType";
+import * as React from "react";
+import { Column as ReactTableColumn, DataRecord, Row as ReactTableRow } from "react-table";
+
+export interface TableProps {
+  /** Table identifier, useful for nested structures */
+  tableId?: string;
+  /** Optional children element to be appended below the table content */
+  children?: React.ReactElement[];
+  /** The prefix to be used for the column name */
+  columnPrefix?: string;
+  /** Optional label to be used for the edit popover that appears when clicking on column header */
+  editColumnLabel?: string;
+  /** For each column there is a default component to be used to render the related cell */
+  defaultCell?: {
+    [columnName: string]: React.FunctionComponent<CellProps>;
+  };
+  /** Table's columns */
+  columns: ReactTableColumn[];
+  /** Table's cells */
+  rows: DataRecord[];
+  /** Function to be executed when columns are modified */
+  onColumnsUpdate?: (columns: ReactTableColumn[]) => void;
+  /** Function to be executed when one or more rows are modified */
+  onRowsUpdate?: (rows: DataRecord[]) => void;
+  /** Function to be executed when a single row gets modified */
+  onSingleRowUpdate?: (rowIndex: number, row: DataRecord) => void;
+  /** Function to be executed when adding a new row to the table */
+  onRowAdding?: () => DataRecord;
+  /** Custom configuration for the table handler */
+  handlerConfiguration: TableHandlerConfiguration;
+  /** The way in which the header will be rendered */
+  headerVisibility?: TableHeaderVisibility;
+  /** Number of levels in the header, 0-based */
+  headerLevels?: number;
+  /** Custom function for getting row key prop, and avoid using the row index */
+  getRowKey?: (row: ReactTableRow) => string;
+  /** Custom function for getting column key prop, and avoid using the column index */
+  getColumnKey?: (column: ReactTableColumn) => string;
+  /** Custom function called for manually resetting a row */
+  resetRowCustomFunction?: (row: DataRecord) => DataRecord;
+}
 
 /** Possible status for the visibility of the Table's Header */
 export enum TableHeaderVisibility {
